Revert "arm64/mpam: fix impossible condition in resctrl_arch_rmid_read()"
This reverts commit fa1a3b356ebb9f6fdb98abd732ac5879ec2713d7.

+++ drivers/platform/mpam/mpam_internal.h
@@ -194,7 +194,7 @@ enum mon_filter_options {
 };
 
 struct mon_cfg {
-	u32                     mon;
+	u16                     mon;
 	u8                      pmg;
 	bool                    match_pmg;
 	u32                     partid;
